+++ api.cpp
@@ -1,0 +1,16 @@
+#include "CustomScript.hpp"
+
+int main() {
+    object obj;
+    obj.flush("Hello");
+
+
+
+}
+void customparser::on_ready(void *) {
+    object object;
+
+    object.flush("Ready!");
+
+
+}