+++ changelog.md
@@ -1,4 +1,23 @@
 # CustomScript Changelog
+### Version 8.6
+#### What's new?
+- Added A JSON Parser! Parse JSON Files from inside CustomScript Now Yay!
+- Added If Parsers Finally! 
+- Added Input parsers
+- Added A **main** function `public parent void* main(String args[]) {}`
+### Version 8.0
+#### What's new?
+- Added new Bash Runtime
+- Added On-the-fly Parsers
+- Added Better functions
+- Added WARNINGS (See MIGRATING.md To see what they are)
+- Added Parser Errors (Finally!)
+- Fixed Problems with block comments
+- Put cache in a ____cscache____ folder
+- Optimized Imports
+- Fixed Buggy keyboard input
+- Added eval() Function
+- Added `mkdir dir(name) ` Function
 ### Version 6.0
 
 #### What's new?
@@ -87,3 +106,12 @@ File defines the filesystem in which files are written, cached, appened, and rea
 ## IOBuff
 This method Is inherited by **iostream** **println** **printf** And **print**.
 It defines the main class In which all console buffers are printed, Deleted, Executed, And Cached.
+## Cache
+This method Is a built-in type.
+Cache Is used by variables using `var`.
+## Delta (Console Buffer)
+Delta Is the elapsed time since last console buffer() method was called. The `del event(<eventname>, <**kwargs>) = {}` Inherits Delta
+## Classdef
+All Classes inherit Classdef, **Keyboard**, **func**, **Function** Inherit classdef.
+## Dir
+Directory. mkdir dir() Functions inherits this class.